Reformatting of file structure, addition of pages, and other changes
Reformatted the file structure so that similar files are grouped.

Added Profile Page, Questions Page, Registration Page, and Change Password Page. Also added routing between pages.

Header was made to change based on whether the user is logged in or not.

Added userData.js file which holds information about the user. This is where the user data that is fetched from the database should be kept.

Began work on a side navigation bar, but could not get it properly implemented yet. Files are there but are not used at the moment.

There are files for a home page, rewards page, user page, and ask question page, but these are placeholders which have not been implemented yet.